fix: starting persistent web socket service (AR-3139) (RC) (#1502)
* fix: start persistent web socket service

* chore: unit test

* chore: revert removing StartServiceReceiver

* chore: cleanup

(cherry picked from commit ddb8d66a64d1f5cb3af1da773a56961bc16351c3)
