+++ app/src/main/AndroidManifest.xml
@@ -160,6 +160,7 @@
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
         </service>
+
         <receiver
                 android:name=".ui.debugscreen.StartServiceReceiver"
                 android:enabled="true"
